manejo de errores, modularizado post y put

+++ app/controllers/itemsApiController.php
@@ -13,9 +13,14 @@ class itemsApiController{
         $this->data = file_get_contents("php://input");
     }
 
+    private function getData() {
+        return json_decode($this->data);
+    }
+
     private function checkGets($queryInfo){//checkea los gets posibles e invalida la peticion en caso de bad request
         $columnasItem = array("id", "nombre", "color", "descripcion");
         $columnasCat = array("especie");
+
 
         if(isset($_GET['order']) && (strtoupper($_GET['order'])=="DESC" || strtoupper($_GET['order'])=="ASC")){
             $queryInfo->order = $_GET['order'];
@@ -23,6 +28,7 @@ class itemsApiController{
             $queryInfo->valido = false;
             $queryInfo->reason = "{Parametro order}";
         }
+
 
         if(isset($_GET['column'])){
             if(in_array(strtolower($_GET['column']), $columnasItem)){
@@ -33,6 +39,22 @@ class itemsApiController{
                 $queryInfo->valido = false;
                 $queryInfo->reason = "{Parametro column}";
             }
+        }
+
+
+        if(isset($_GET['offset']) && is_numeric($_GET['offset'])){
+            $queryInfo->offset = $_GET['offset'];
+        }else if(isset($_GET['offset']) && !is_numeric($_GET['offset'])){
+            $queryInfo->valido = false;
+            $queryInfo->reason = "{Parametro offset}";
+        }
+
+
+        if(isset($_GET['lenght']) && is_numeric($_GET['lenght'])){
+            $queryInfo->lenght = $_GET['lenght'];
+        }else if(isset($_GET['lenght']) && !is_numeric($_GET['lenght'])){
+            $queryInfo->valido = false;
+            $queryInfo->reason = "{Parametro lenght}";
         }
     }
 
@@ -82,6 +104,8 @@ class itemsApiController{
         $queryInfo->column = "raza.nombre";
         $queryInfo->order = "ASC";
         $queryInfo->reason = "valid";
+        $queryInfo->offset = 0;
+        $queryInfo->lenght = null;
 
         if(isset($params[':ID'])){//uno solo
             $item = $this->getOne($params, $queryInfo);
@@ -91,9 +115,10 @@ class itemsApiController{
             $item = $this->getGroup($especie, $queryInfo);
 
         }else if(empty($params)){//trae todos los animales
+            
             $this->checkGets($queryInfo);
             if($queryInfo->valido){    
-                $item = $this->model->getAllItems($queryInfo->column, $queryInfo->order);
+                $item = $this->model->getAllItems($queryInfo->column, $queryInfo->order, $queryInfo->offset, $queryInfo->lenght);
             }
         }
 
@@ -101,44 +126,53 @@ class itemsApiController{
     }
 
 
-
-
-
-
-
-
-
-
-
-
-
-
-
     public function deleteItem($params = []){
         $id = $params[':ID'];
-        $item = $this->model->getOneItem($id);
-        if (!empty($item)) {
-            $this->model->deleteItem($id);
+
+        if(is_numeric($id)){
             $item = $this->model->getOneItem($id);
-            if(!$item){
-                $this->view->response("Item id=$id eliminado con éxito", 200);
+            if (!empty($item)) {
+                $this->model->deleteItem($id);
+                $item = $this->model->getOneItem($id);
+                if(!$item){
+                    $this->view->response("Item id=$id eliminado con éxito", 200);
+                }else{
+                    $this->view->response("El servidor no pudo borrar por una falla interna", 500);
+                }
             }
-        }
-        else
-        $this->view->response("item id=$id not found", 404);
+            else
+            $this->view->response("item id=$id not found", 404);
+        }else
+        $this->view->response("Solo se aceptan id numericos", 400);
+
+        
     }
+
+
+
+
+
+    private function checkVoidInputs($body){
+        return !(empty($body->nombre) || empty($body->color) || empty($body->descripcion) || empty($body->especie) || ctype_space($body->nombre) || ctype_space($body->color) || ctype_space($body->descripcion) || ctype_space($body->especie));
+    }
+
+
+
+
 
     public function post($params = []){
         $body = $this->getData();
 
-        if(empty($body->nombre) || empty($body->color) || empty($body->descripcion) || empty($body->especie) || ctype_space($body->nombre) || ctype_space($body->color) || ctype_space($body->descripcion) || ctype_space($body->especie)){
+        if($this->checkVoidInputs($body)){
             $this->view->response("Complete los datos", 400);
         }
         else{
             $category = $this->model->getCat($body->especie);
             if(!empty($category)){
-                $id = $this->model->addItem($body->nombre, $body->color, $body->descripcion, 
-                $body->especie);
+                $id = $this->model->addItem($body->nombre, 
+                                            $body->color, 
+                                            $body->descripcion, 
+                                            $body->especie);
 
                 $item = $this->model->getOneItem($id);
                 $this->view->response($item, 201);
@@ -148,32 +182,39 @@ class itemsApiController{
         }
     }
 
+
+
+
+
+
     public function put($params = []){
         $id = $params[':ID'];
         $item = $this->model->getOneItem($id);
 
-        if (!empty($item)) {//se fija que exista el item a modificar
+        if (!empty($item) && is_numeric($id)) {//se fija que exista el item a modificar
             $body = $this->getData();
 
-            if(empty($body->nombre) || empty($body->color) || empty($body->descripcion) || empty($body->especie) || ctype_space($body->nombre) || ctype_space($body->color) || ctype_space($body->descripcion) || ctype_space($body->especie)){//los campos deben estar completos
+            if($this->checkVoidInputs($body)){//los campos deben estar completos
                 $this->view->response("Complete los datos", 400);
             }else{
                 $category = $this->model->getCat($body->especie);
-                if(!empty($category)){
-                    $this->model->modItem($body->nombre, $body->color,                 $body->descripcion, $body->especie, $id);
+                if(!empty($category)){//tiene que existir la categoria
+                    $this->model->modItem($body->nombre, 
+                                            $body->color, 
+                                            $body->descripcion, 
+                                            $body->especie, 
+                                            $id);
                     $this->view->response($this->model->getOneItem($id), 200);
                 }
                 else
                 $this->view->response("No existe la categoria con id=$body->especie", 400);
             }
+
+        }else if(!is_numeric($id)){
+            $this->view->response("Solo se aceptan id numericos", 400);
         }
         else
         $this->view->response("item id=$id not found", 404);
        
     }
-
-    private function getData() {
-        return json_decode($this->data);
-    }
-
 }
